RecordIDs on the PalmOS handhelds are unsigned long (highest 3 bits or so are unused), so don't use int, but really unsigned long instead.
svn path=/trunk/kdepim/; revision=332500

+++ libkcal/incidencebase.cpp
@@ -335,14 +335,14 @@ int IncidenceBase::syncStatus() const
   return mSyncStatus;
 }
 
-void IncidenceBase::setPilotId( int id )
+void IncidenceBase::setPilotId( unsigned long id )
 {
   if (mReadOnly) return;
 
   mPilotId = id;
 }
 
-int IncidenceBase::pilotId() const
+unsigned long IncidenceBase::pilotId() const
 {
   return mPilotId;
 }
